+++ src/components/User/Menu/__tests__/Menu.test.tsx
@@ -1,8 +1,8 @@
 import { render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import userEvent from '@testing-library/user-event';
-import { renderWithTheme } from '../../../utils/theme';
-import Menu from '.';
+import { renderWithTheme } from '../../../../utils/theme';
+import Menu from '..';
 
 describe('<Menu />', () => {
   const props = {
