+++ src/components/User/SectionItem/index.tsx
@@ -3,8 +3,8 @@ import { styled } from '@mui/material/styles';
 import Stack from '@mui/material/Stack';
 import Typography from '@mui/material/Typography';
 import Avatar from '@mui/material/Avatar';
-import TextTag from '../../shared/TextTag/TextTag';
-import Link from '../../shared/Link/Link';
+import TextTag from '../../shared/TextTag';
+import Link from '../../shared/Link';
 import { Repo, User, SectionType } from '../../../utils/types';
 
 const Root = styled('div')(({ theme }) => ({
